requirements.txt
python-dotenv
rich
typer
openai
yfinance
schedule

